# parent/com.bike.server/target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: jacquetq@gmail.com
  description: |
    Bike Rest Api with Spring Boot
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Bike Api
  version: 1.0.0
...
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/QuentinJacquet/com.bike.api/1.0.0
...
paths:
...
  /ride:
    get:
      operationId: getRides
      parameters:
      - description: ID of user
        explode: true
        in: query
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - description: filter rides by status
        explode: true
        in: query
        name: completed
        required: false
        schema:
          type: boolean
        style: form
      - description: filter rides begun after given date
        explode: true
        in: query
        name: startDate
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - description: filter rides ended before given date
        explode: true
        in: query
        name: endDate
        required: false
        schema:
          format: date-time
          type: string
        style: form
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Ride'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Ride'
                type: array
          description: successful operation
      summary: Get all rides
      tags:
      - ride
      x-accepts: application/json
      x-tags:
      - tag: ride
  /ride/start:
    post:
      operationId: startRide
      parameters:
      - description: ID of user
        explode: true
        in: query
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - description: ID of station
        explode: true
        in: query
        name: stationId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      requestBody:
        $ref: '#/components/requestBodies/Ride'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ride'
            application/xml:
              schema:
                $ref: '#/components/schemas/Ride'
          description: successful operation
        "405":
          description: Invalid input
      summary: start a new ride
      tags:
      - ride
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: ride
  /ride/{rideId}:
    delete:
      operationId: deleteRide
      parameters:
      - description: Ride id to delete
        explode: false
        in: path
        name: rideId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Ride not found
      summary: Deletes a ride
      tags:
      - ride
      x-accepts: application/json
      x-tags:
      - tag: ride
    get:
      description: Returns a single ride
      operationId: getRideById
      parameters:
      - description: ID of ride to return
        explode: false
        in: path
        name: rideId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ride'
            application/xml:
              schema:
                $ref: '#/components/schemas/Ride'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Ride not found
      summary: Find ride by ID
      tags:
      - ride
      x-accepts: application/json
      x-tags:
      - tag: ride
  /ride/{rideId}/end:
    post:
      operationId: endRide
      parameters:
      - description: ID of ride to end
        explode: false
        in: path
        name: rideId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: ID of station
        explode: true
        in: query
        name: stationId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Ride not found
      summary: End a ride
      tags:
      - ride
      x-accepts: application/json
      x-tags:
      - tag: ride
components:
  requestBodies:
    Ride:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Ride'
        application/xml:
          schema:
            $ref: '#/components/schemas/Ride'
      description: Ride object that needs to be added
      required: true
...
  schemas:
...
    Ride:
      example:
        endDate: 2000-01-23T04:56:07.000+00:00
        endStationId: 5
        id: 0
        startStationId: 1
        complete: false
        userId: 6
        startDate: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          format: int64
          readOnly: true
          type: integer
        userId:
          format: int64
          type: integer
        startStationId:
          format: int64
          type: integer
        startDate:
          format: date-time
          type: string
        endStationId:
          format: int64
          type: integer
        endDate:
          format: date-time
          type: string
        complete:
          default: false
          type: boolean
      required:
      - startDate
      - startStationId
      - userId
      type: object
      xml:
        name: Ride
...